# MediaServer/docs/swagger.yaml
definitions:
  AuthHandler.UserCredential:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  collection.CollectionUpsertDelFormat:
    properties:
      image_uuids:
        items:
          type: string
        type: array
    type: object
  collection.NewCollection:
    properties:
      description:
        type: string
      name:
        type: string
      thumbnail_uuid:
        type: string
    type: object
  mediahandler.DeleteRequest:
    properties:
      delete_id:
        items:
          type: string
        type: array
    type: object
  serverutils.PaginatedResponse:
    properties:
      content:
        items:
          additionalProperties: {}
          type: object
        type: array
      next:
        type: string
    type: object
info:
  contact: {}
  description: API for interacting with the media server
  title: Media Server API
  version: "0.1"
paths:
  /auth/me:
    get:
      description: Returns the authenticated user's UUID, username, and account creation
        timestamp. Requires a valid auth_token cookie.
      produces:
      - application/json
      responses:
        "200":
          description: uuid, username, created_at
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get current user info
      tags:
      - Auth
  /collection:
    post:
      consumes:
      - application/json
      description: Creates a new collection owned by the authenticated user. Returns
        201 with the new collection UUID as plain text.
      parameters:
      - description: Collection name, description, and optional thumbnail UUID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/collection.NewCollection'
      produces:
      - text/plain
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create a new collection
      tags:
      - Creation
  /collection/{collection_id}:
    delete:
      description: Permanently deletes a collection owned by the authenticated user.
        Requires owner-level permission. Returns 204 on success.
      parameters:
      - description: UUID of the collection
        in: path
        name: collection_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a collection
      tags:
      - Deletion
    get:
      description: Returns full metadata for a single collection the user has access
        to. Returns 403 if the user lacks permission (requires at least basicviewerpermlevel).
      parameters:
      - description: UUID of the collection
        in: path
        name: collection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Array of collection row maps
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get collection info by ID
      tags:
      - Query
    post:
      consumes:
      - application/json
      description: Adds one or more images (by UUID) to an existing collection. The
        caller must own the images and have content-editor permission on the collection.
        Returns 201 on success.
      parameters:
      - description: UUID of the collection
        in: path
        name: collection_id
        required: true
        type: string
      - description: Array of image UUIDs to insert
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/collection.CollectionUpsertDelFormat'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      summary: Insert images into a collection
      tags:
      - Creation
  /collection/{collection_id}/contents:
    get:
      description: Returns a paginated list of images belonging to the specified collection,
        ordered by date added. Supports `want`, `offset`, and `limit` query parameters.
        The `want` parameter is a dash-separated list of fields (e.g. `uuid-original_url-thumbnail_url-preview_url-created_at`).
        Requires content-viewer permission on the collection.
      parameters:
      - description: UUID of the collection
        in: path
        name: collection_id
        required: true
        type: string
      - description: Dash-separated list of fields to return (e.g. uuid-original_url-thumbnail_url)
        in: query
        name: want
        type: string
      - description: Number of entries to skip
        in: query
        name: offset
        type: integer
      - description: Maximum number of entries to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated content array with optional next URL
          schema:
            $ref: '#/definitions/serverutils.PaginatedResponse'
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get collection image contents
      tags:
      - Query
  /gms/collection:
    get:
      description: Returns a paginated list of all collections the authenticated user
        has any level of access to, ordered by creation date descending.
      parameters:
      - description: Number of entries to skip
        in: query
        name: offset
        type: integer
      - description: Maximum number of entries to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: content array of { uuid, created_at, name, description, thumbnail_url
            }
          schema:
            $ref: '#/definitions/serverutils.PaginatedResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: List available collections
      tags:
      - Query
  /gms/media:
    get:
      description: Query for media available to a user
      parameters:
      - description: Offset from the first of an ordered list
        in: query
        name: offset
        type: integer
      - description: Limit the number of entries returned
        in: query
        name: limit
        type: integer
      - description: Which order is it in
        in: query
        name: order
        type: string
      - description: Which attribute are the entries ordered by
        in: query
        name: by
        type: string
      - description: Which attributes are wanted in the response
        in: query
        name: want
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Query for media test for changs
      tags:
      - Query
    post:
      consumes:
      - multipart/form-data
      description: Upload images via form data in the field "files"
      produces:
      - text/plain
      responses:
        "200":
          description: OK
      summary: Upload images
      tags:
      - Creation
  /gms/media/{type}/{id}:
    get:
      description: Retrieve a media file by its UUID and type. Returns the file via
        X-Accel-Redirect (served by nginx). Returns 403 if the authenticated user
        does not own the media.
      parameters:
      - description: Media type (e.g. thumbnail, original)
        in: path
        name: type
        required: true
        type: string
      - description: UUID of the media item
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: Get media file by ID
      tags:
      - Query
  /gms/media/query:
    post:
      description: Query for by word or filter
      parameters:
      - description: Text to search for
        in: query
        name: text_query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Query media
      tags:
      - Query
  /gms/media/suggestions:
    get:
      description: Query for media similar to a specified uuid
      parameters:
      - description: uuid of the root media
        in: query
        name: uuid
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get media suggestions
      tags:
      - Query
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate with username and password. Sets an HttpOnly cookie
        named auth_token (JWT, 7-day expiry) on success and returns the user UUID.
      parameters:
      - description: Username and password
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/AuthHandler.UserCredential'
      produces:
      - application/json
      responses:
        "200":
          description: userid field contains the user UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Log in
      tags:
      - Auth
  /media/delete:
    post:
      consumes:
      - application/json
      description: Delete one or more images owned by the authenticated user. Only
        images whose owner_uuid matches the caller are deleted. Returns 204 on success.
      parameters:
      - description: List of image UUIDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/mediahandler.DeleteRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete media by ID
      tags:
      - Deletion
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user account with a username and password. Also creates
        a default collection for the user. Returns 201 on success.
      parameters:
      - description: Username and password
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/AuthHandler.UserCredential'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Register a new user
      tags:
      - Auth
swagger: "2.0"